Remove initial pageview send on insert
On its own this plugin does not reliably send pageview/transition events to GA. To fix you still have to add an observer to the router to send the pageview events. If you do that then sending the event when injecting the GA code causes a duplicate initial event. So this commit removes that pageview event when injecting the tracking code into the html.